+++ C971_MobileApp/C971_MobileApp/MainPage.xaml.cs
@@ -33,11 +33,15 @@ namespace C971_MobileApp
 
             Navigation.PushAsync(new EditCourse(buttonId));
         }
-        public void DeleteCourse(object sender, EventArgs e)
+        public async void DeleteCourse(object sender, EventArgs e)
         {
-            ImageButton button = (ImageButton)sender;
-            App.CourseData.DeleteCourse((int)button.BindingContext);
-            RefreshCourses();
+            bool deleteCourse = await DisplayAlert("Deleting Course", "Are you sure you want to delete this course?", "Yes", "No");
+            if (deleteCourse)
+            {
+                ImageButton button = (ImageButton)sender;
+                App.CourseData.DeleteCourse((int)button.BindingContext);
+                RefreshCourses();
+            }
         }
         private void RefreshCourses()
         {
